requirements.txt
decord
dm-tree
ffmpeg-python
ffmpy
librosa
matplotlib
moviepy
opencv-python
pydub
PyYAML
scikit-learn
timm==0.4.5
torch
torchaudio
tqdm
x_transformers==0.27.1
gym==0.19
gym3
attrs
gdown
accelerate==0.18.0 
wandb
importlib-resources==5.12
